Add review tests

+++ app/Http/Controllers/ReviewController.php
@@ -2,46 +2,36 @@
 
 namespace App\Http\Controllers;
 
+use Throwable;
 use App\Models\Review;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use App\Http\Requests\ReviewRequest;
 use App\Http\Resources\ReviewResource;
 use Symfony\Component\HttpFoundation\Response;
+use App\Exceptions\ProductNotBelongsToUserException;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
 class ReviewController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth:api')->only(['store', 'update', 'destroy']);
+        $this->middleware('auth:api')
+            ->only(['store', 'update', 'destroy']);
     }
 
     /**
      * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
      */
-    public function index(Product $product)
+    public function index(Product $product): AnonymousResourceCollection
     {
         return ReviewResource::collection($product->reviews);
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
      * Store a newly created resource in storage.
-     *
-     * @return Response
      */
-    public function store(ReviewRequest $request, Product $product)
+    public function store(ReviewRequest $request, Product $product): Response
     {
         $review = new Review($request->all());
 
@@ -53,46 +43,26 @@ class ReviewController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Review $review)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Review $review)
-    {
-        //
-    }
-
-    /**
      * Update the specified resource in storage.
-     *
-     * @return Response
      */
-    public function update(Request $request, Product $product, Review $review)
+    public function update(Request $request, Product $product, Review $review): Response
     {
         $review->update($request->all());
 
         return response([
             'data' => new ReviewResource($review),
-        ], Response::HTTP_CREATED);
+        ], Response::HTTP_OK);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @return Response
+     * @throws Throwable
      */
-    public function destroy(Product $product, Review $review)
+    public function destroy(Request $request, Product $product, Review $review): Response
     {
+        throw_if($request->user()->id !== $product->user_id, new ProductNotBelongsToUserException);
+
         $review->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
